Refactor Linux service handling into shared modules
- Move Linux CLI socket diagnostics and startup checks into `linux.rs`
- Centralize daemon socket permission setup in `playitd::linux`
- Preserve clearer IPC errors for installed service startup
